Update PCA plots

+++ Rank_rank_prep.py
@@ -14,8 +14,8 @@ def generate_rrho_file(input_file, wd):
     for i in range(len(columns_)):
         for j in range(i+1, len(columns_)):
             rrho_file = pd.DataFrame({"Unigene":"N/A", "Gene":rf.index.tolist(), columns_[j]:rf[columns_[j]], columns_[i]:rf[columns_[i]]})
-            j_rank = rrho_file[columns_[j]].rank(ascending=False)
-            i_rank = rrho_file[columns_[i]].rank(ascending=False)
+            j_rank = rrho_file[columns_[j]].rank(ascending=False).astype(int)
+            i_rank = rrho_file[columns_[i]].rank(ascending=False).astype(int)
             rrho_file.insert(2, columns_[j]+"_rank", j_rank)
             rrho_file.insert(3, columns_[i]+"_rank", i_rank)
             rrho_file = rrho_file[["Unigene", "Gene", columns_[j]+"_rank", columns_[i]+"_rank", columns_[j], columns_[i]]]
